Untested implementation of sdt and load literal

+++ src/sdtInstruction.h
@@ -1,5 +1,5 @@
 #ifndef SDT_INSTRUCTION_H
 #define SDT_INSTRUCTION_H
 
-void executeSdtInstruction(ComputerState*, const int64_t instruction);
+void executeSdtInstruction(ComputerState*, const int32_t instruction);
 #endif // SDT_INSTRUCTION_H
